\name{checkdifferentiation}
\alias{checkdifferentiation}
\title{Check Differentiation}
\usage{
checkdifferentiation(heatmap, habitats)
}
\description{
Calculates the spatial isolation statistic for every generation
}
\examples{
checkdifferentiation(heatmap = "heatmap.csv", habitats = "habitat.csv")
}
\author{Joris Damhuis}
