\name{checkspeciation}
\alias{checkspeciation}
\title{Check Speciation}
\usage{
checkspeciation(heatmap, filename, method, output = NA, calledbyfunction = FALSE)
}
\description{
Calculates the ecological speciation statistic for every generation
}
\examples{
checkspeciation(heatmap = "heatmap.csv", filename = "test.csv", method = "completion", output = "output.csv", calledbyfunction = FALSE)
}
\author{Joris Damhuis}
