\name{makeplots}
\alias{makeplots}
\title{Create all useful plots}
\usage{
makeplots(filename, heatmap, stepsize = 0.05, everyntimesteps = 1, habitat = NA)
}
\description{
Creates all useful plots to analyse the behaviour of the simulation
}
\examples{
makeplots(filename = "test.csv", heatmap = "heatmap.csv", stepsize = 0.05, everyntimesteps = 1, habitat = "habitat.csv")
}
\author{Joris Damhuis}
