\name{outcomeplot}
\alias{outcomeplot}
\title{Create the outcome plot}
\usage{
outcomeplot(filename, nsim, title = NA)
}
\description{
Creates a heatmap-like plot to summarise the outcomes multiple simulations
}
\examples{
outcomeplot(filename = "Delta0.2New.csv", nsim = 10, title = "Delta = 0.2")
}
\author{Joris Damhuis}
